+++ Primitive_Drawing/EventHandler.cpp
@@ -15,7 +15,7 @@ void  EventHandler::handleKeyPress(GLFWwindow* window, int key, int scancode, in
 void EventHandler::handleMouseMove(GLFWwindow* window, double posx, double posy) {
 	vec2 mousePosition = applicationManger->getMousePosition();
 	vec2 screenResolution = applicationManger->getscreenResolution();
-	applicationManger->getRenderer()->handleMouseMove(-posx + mousePosition.x, posy - mousePosition.y);
+	applicationManger->getRenderer()->handleMouseMove((float)(-posx + mousePosition.x), (float)(posy - mousePosition.y));
 	applicationManger->resetMousePosition();
 }
