+++ Primitive_Drawing/TPCamera.h
@@ -17,7 +17,7 @@ public:
 	void Fly(float distance);
 
 	void UpdateViewMatrix();
-	void Update(double time);
+	void Update(float time);
 	void setDistance(float distance);
 
 	~TPCamera();
